Refactor: Rename residentCard to idCard
