--- src/modules/ResidentCardCreation/Form/index.ts
@@ -1,1 +1,0 @@
-export * from './ResidentCardForm';
